+++ apps/E2E/src/Callout/pages/CalloutPageObject.win.ts
@@ -1,9 +1,4 @@
-import {
-  CALLOUT_TESTPAGE,
-  CALLOUT_TEST_COMPONENT,
-  HOMEPAGE_CALLOUT_BUTTON,
-  BUTTON_TO_OPEN_CALLOUT,
-} from '../consts';
+import { CALLOUT_TESTPAGE, CALLOUT_TEST_COMPONENT, HOMEPAGE_CALLOUT_BUTTON, BUTTON_TO_OPEN_CALLOUT } from '../consts';
 import { BasePage, By } from '../../common/BasePage';
 
 class CalloutPageObject extends BasePage {
@@ -33,20 +28,16 @@ class CalloutPageObject extends BasePage {
   /*****************************************/
   /**************** Getters ****************/
   /*****************************************/
-  get _testPage() {
-    return By(CALLOUT_TESTPAGE);
-  }
-
   get _pageName() {
     return CALLOUT_TESTPAGE;
   }
 
-  get _primaryComponent() {
-    return By(CALLOUT_TEST_COMPONENT);
+  get _primaryComponentName() {
+    return CALLOUT_TEST_COMPONENT;
   }
 
-  get _pageButton() {
-    return By(HOMEPAGE_CALLOUT_BUTTON);
+  get _pageButtonName() {
+    return HOMEPAGE_CALLOUT_BUTTON;
   }
 
   get _buttonToOpenCallout() {