+++ apps/E2E/src/Pressable/pages/PressablePageObject.ts
@@ -1,28 +1,20 @@
-import {
-  PRESSABLE_TESTPAGE,
-  PRESSABLE_TEST_COMPONENT,
-  HOMEPAGE_PRESSABLE_BUTTON,
-} from '../consts';
-import { BasePage, By } from '../../common/BasePage';
+import { PRESSABLE_TESTPAGE, PRESSABLE_TEST_COMPONENT, HOMEPAGE_PRESSABLE_BUTTON } from '../consts';
+import { BasePage } from '../../common/BasePage';
 
 class PressablePageObject extends BasePage {
   /*****************************************/
   /**************** Getters ****************/
   /*****************************************/
-  get _testPage() {
-    return By(PRESSABLE_TESTPAGE);
-  }
-
   get _pageName() {
     return PRESSABLE_TESTPAGE;
   }
 
-  get _primaryComponent() {
-    return By(PRESSABLE_TEST_COMPONENT);
+  get _primaryComponentName() {
+    return PRESSABLE_TEST_COMPONENT;
   }
 
-  get _pageButton() {
-    return By(HOMEPAGE_PRESSABLE_BUTTON);
+  get _pageButtonName() {
+    return HOMEPAGE_PRESSABLE_BUTTON;
   }
 }