+++ apps/E2E/src/Shimmer/pages/ShimmerPageObject.win.ts
@@ -1,28 +1,20 @@
-import {
-  HOMEPAGE_SHIMMER_BUTTON,
-  SHIMMER_TESTPAGE,
-  SHIMMER_TEST_COMPONENT,
-} from '../consts';
-import { BasePage, By } from '../../common/BasePage';
+import { HOMEPAGE_SHIMMER_BUTTON, SHIMMER_TESTPAGE, SHIMMER_TEST_COMPONENT } from '../consts';
+import { BasePage } from '../../common/BasePage';
 
 class ShimmerPageObject extends BasePage {
   /*****************************************/
   /**************** Getters ****************/
   /*****************************************/
-  get _testPage() {
-    return By(SHIMMER_TESTPAGE);
-  }
-
   get _pageName() {
     return SHIMMER_TESTPAGE;
   }
 
-  get _primaryComponent() {
-    return By(SHIMMER_TEST_COMPONENT);
+  get _primaryComponentName() {
+    return SHIMMER_TEST_COMPONENT;
   }
 
-  get _pageButton() {
-    return By(HOMEPAGE_SHIMMER_BUTTON);
+  get _pageButtonName() {
+    return HOMEPAGE_SHIMMER_BUTTON;
   }
 }