+++ packages/experimental/RadioGroup/src/Radio/RadioTokens.ts
@@ -5,6 +5,7 @@ import { RadioTokens } from './Radio.types';
 
 export const defaultRadioTokens: TokenSettings<RadioTokens, Theme> = (t: Theme) =>
   ({
+    rootHorizontalPadding: globalTokens.size40,
     borderColor: t.colors.transparentStroke,
     borderStyle: 'solid',
     borderWidth: globalTokens.stroke.width10,
@@ -15,7 +16,7 @@ export const defaultRadioTokens: TokenSettings<RadioTokens, Theme> = (t: Theme) 
     radioBorderStyle: 'solid',
     radioBorder: t.colors.neutralStrokeAccessible,
     color: t.colors.neutralForeground3,
-    radioSize: 20,
+    radioOuterCircleSize: 20,
     radioInnerCircleSize: 10,
     labelMarginTop: globalTokens.size40 + 1,
     labelMarginRight: globalTokens.size20,
@@ -29,11 +30,14 @@ export const defaultRadioTokens: TokenSettings<RadioTokens, Theme> = (t: Theme) 
     marginLeft: globalTokens.size40,
     flexDirection: 'row',
     alignItems: 'flex-start',
+    labelAlignItems: 'flex-start',
 
     labelPositionBelow: {
       flexDirection: 'column',
       alignItems: 'center',
-      marginLeft: globalTokens.sizeNone,
+      labelAlignItems: 'center',
+      labelMarginLeft: globalTokens.size20,
+      marginLeft: globalTokens.size60,
     },
 
     disabled: {