+++ packages/experimental/RadioGroup/src/Radio/useRadio.ts
@@ -102,6 +102,7 @@ export const useRadio = (props: RadioProps): RadioInfo => {
       accessibilityPositionInSet: accessibilityPositionInSet ?? radioGroupContext.values.findIndex((x) => x == value) + 1,
       accessibilitySetSize: accessibilitySetSize ?? radioGroupContext.values.length,
       focusable: !state.disabled,
+      disabled: isDisabled,
       enableFocusRing: enableFocusRing ?? true,
       onAccessibilityAction: onAccessibilityAction,
     },