+++ packages/experimental/Shadow/CHANGELOG.md
@@ -1,8 +1,37 @@
 # Change Log - @fluentui-react-native/experimental-shadow
 
-This log was last generated on Fri, 06 Jan 2023 12:26:31 GMT and should not be manually modified.
+This log was last generated on Wed, 18 Jan 2023 01:55:11 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.2.60
+
+Wed, 18 Jan 2023 01:55:11 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/framework to v0.8.34
+- Bump @fluentui-react-native/pressable to v0.9.50
+- Bump @fluentui-react-native/theme-types to v0.29.1
+
+## 0.2.59
+
+Tue, 10 Jan 2023 20:17:18 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/framework to v0.8.33
+- Bump @fluentui-react-native/pressable to v0.9.49
+- Bump @fluentui-react-native/theme-types to v0.29.0
+
+## 0.2.58
+
+Mon, 09 Jan 2023 22:25:00 GMT
+
+### Patches
+
+- Bump @fluentui-react-native/framework to v0.8.32
+- Bump @fluentui-react-native/pressable to v0.9.48
 
 ## 0.2.57